requirements.txt
torch
torchvision
tqdm
numpy
scipy
scikit-learn
pandas
pydantic
opencv-python
pyyaml
tabulate
numba
typing_extensions
wandb
plotly
python-dotenv
h5py
pytorch-lightning
tensorflow_probability
